MXS-3613: Add support for metadata skip
The MARIADB_CLIENT_CACHE_METADATA flag is now advertised by MaxScale and
understood by the result parsing code. Added the
RCAP_TYPE_NO_METADATA_CACHE routing flag to allow the feature to be
disabled for modules that don't support the caching of metadata. Currently
cache, masking and smartrouter modules do not support it as they use
different code for processing the result metadata.